infra/pyproject.toml
[tool.poetry]
name = "infra"
version = "0.1.0"
description = ""
authors = ["Masaki Kobayashi <bayashi.cl@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "2.150.0"
constructs = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
